+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # Changelog
 
 All notable changes to `laravel-ugc-translate` will be documented in this file.
+
+## [0.2.0] - 2024-12-10
+
+Added optional google translate service, improved documentation
+
+### Added
+- [[FEATURE-ADDING-GOOGLE-TRANSLATE]](https://github.com/rpwebdevelopment/laravel-ugc-translate/tree/feature-adding-google-translate)
+
+### Changed
+
+### Fixed